requirements.txt
elevenlabs==2.6.0
faiss_cpu==1.11.0
langchain==0.3.26
langchain_community==0.3.27
langchain_core==0.3.68
langchain_ollama==0.3.3
openai_whisper==20250625
python-dotenv==1.1.1
sounddevice==0.5.2
soundfile==0.13.1
streamlit==1.46.0
